Fixed issues as requested
Splited the login into 3 files, one for each class
Moved the method check board to the class board
